pyproject.toml
[project]
name = "pyvcontrol"
version = "1.5.0"
description = "Communication with Viessmann heating via Optolink interface."
readme = "README.md"
authors = [
    { name = "Jochen Schmähling", email = "j_ochen_573@freenet.de" }
]
requires-python = ">=3.12"
dependencies = [
    "deprecated>=1.3.1",
    "pyserial>=3.5",
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
